backend/requirements.txt
# backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0
cryptography>=42.0,<43.0
google-generativeai>=0.5,<1.0
requests>=2.31,<3.0
Pillow>=10.0,<11.0
boto3>=1.34,<2.0
django-storages>=1.14,<2.0
python-dotenv>=1.0,<2.0
gunicorn>=22.0,<23.0

# Testing
pytest>=8.0,<10.0
pytest-django>=4.8,<5.0
